[#1965] Use regular CriteriaBuilder instead of PaginatedCriteriaBuilder for unpaged Spring Data JPA repository invocations
