+++ integration/spring-data/base/src/main/java/com/blazebit/persistence/spring/data/base/repository/AbstractEntityViewAwareRepository.java
@@ -15,6 +15,7 @@ import com.blazebit.persistence.criteria.BlazeCriteria;
 import com.blazebit.persistence.parser.EntityMetamodel;
 import com.blazebit.persistence.spi.ExtendedManagedType;
 import com.blazebit.persistence.spring.data.base.EntityViewSortUtil;
+import com.blazebit.persistence.spring.data.base.PageableUtil;
 import com.blazebit.persistence.spring.data.base.query.KeysetAwarePageImpl;
 import com.blazebit.persistence.spring.data.repository.KeysetPageable;
 import com.blazebit.persistence.view.EntityViewManager;
@@ -69,18 +70,6 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
     private static final String DELETE_ALL_QUERY_BY_ID_STRING = "delete from %s x where %s in :ids";
     private static final String[] EMPTY = new String[0];
     private static final EscapeCharacter DEFAULT = EscapeCharacter.of('\\');
-    private static final Pageable UNPAGED;
-
-    static {
-        Pageable unpaged = null;
-        try {
-            Method unpagedMethod = Class.forName("org.springframework.data.domain.Pageable").getMethod("unpaged");
-            unpaged = (Pageable) unpagedMethod.invoke(null);
-        } catch (Exception e) {
-            // ignore
-        }
-        UNPAGED = unpaged;
-    }
 
     protected EscapeCharacter escapeCharacter = DEFAULT;
 
@@ -331,7 +320,7 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
         Class<S> probeType = example.getProbeType();
         TypedQuery<S> query = getQuery(new ExampleSpecification<>(example, escapeCharacter), probeType, pageable);
 
-        return pageable == null || pageable == UNPAGED ? new KeysetAwarePageImpl<>(query.getResultList()) : new KeysetAwarePageImpl<>((PagedList<S>) query.getResultList(), pageable);
+        return PageableUtil.isUnpaged(pageable) ? new KeysetAwarePageImpl<>(query.getResultList()) : new KeysetAwarePageImpl<>((PagedList<S>) query.getResultList(), pageable);
     }
 
     public List<E> findAll(Sort sort) {
@@ -339,7 +328,7 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
     }
 
     public Page<E> findAll(Pageable pageable) {
-        if (pageable == null || pageable == UNPAGED) {
+        if (PageableUtil.isUnpaged(pageable)) {
             return (Page<E>) new PageImpl<>(findAll());
         }
 
@@ -502,7 +491,7 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
 
     public Page<V> findAll(Specification<E> spec, Pageable pageable) {
         TypedQuery<V> query = getQuery(spec, pageable);
-        if (pageable == null || pageable == UNPAGED) {
+        if (PageableUtil.isUnpaged(pageable)) {
             return new KeysetAwarePageImpl<>(query.getResultList());
         }
         PagedList<V> resultList = (PagedList<V>) query.getResultList();
@@ -564,7 +553,7 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
 
         TypedQuery<V> query;
         if (entityViewClass == null) {
-            if (pageable == null || pageable == UNPAGED) {
+            if (PageableUtil.isUnpaged(pageable)) {
                 query = (TypedQuery<V>) cb.getQuery();
             } else {
                 PaginatedCriteriaBuilder<S> paginatedCriteriaBuilder;
@@ -585,7 +574,7 @@ public abstract class AbstractEntityViewAwareRepository<V, E, ID extends Seriali
                 query = (TypedQuery<V>) paginatedCriteriaBuilder.getQuery();
             }
         } else {
-            if (pageable == null || pageable == UNPAGED) {
+            if (PageableUtil.isUnpaged(pageable)) {
                 EntityViewSetting<V, CriteriaBuilder<V>> setting = EntityViewSetting.create(entityViewClass);
                 CriteriaBuilder<V> fqb = evm.applySetting(setting, cb);
                 if (sort != null) {
