+++ integration/spring-data/testsuite/webmvc/src/test/java/com/blazebit/persistence/spring/data/testsuite/webmvc/view/DocumentView.java
@@ -6,6 +6,7 @@
 package com.blazebit.persistence.spring.data.testsuite.webmvc.view;
 
 import com.blazebit.persistence.spring.data.testsuite.webmvc.entity.Document;
+import com.blazebit.persistence.spring.data.testsuite.webmvc.entity.Identifiable;
 import com.blazebit.persistence.view.EntityView;
 import com.blazebit.persistence.view.IdMapping;
 import com.blazebit.persistence.view.Mapping;
@@ -18,9 +19,10 @@ import java.time.Instant;
  * @since 1.5.0
  */
 @EntityView(Document.class)
-public interface DocumentView {
+public interface DocumentView extends Identifiable<Long> {
 
     @IdMapping
+    @Override
     Long getId();
 
     String getName();
