+++ integration/spring-data/testsuite/webmvc/src/test/java/com/blazebit/persistence/spring/data/testsuite/webmvc/view/PersonUpdateView.java
@@ -5,6 +5,7 @@
 
 package com.blazebit.persistence.spring.data.testsuite.webmvc.view;
 
+import com.blazebit.persistence.spring.data.testsuite.webmvc.entity.Identifiable;
 import com.blazebit.persistence.spring.data.testsuite.webmvc.entity.Person;
 import com.blazebit.persistence.view.EntityView;
 import com.blazebit.persistence.view.IdMapping;
@@ -18,9 +19,10 @@ import java.util.UUID;
  */
 @UpdatableEntityView
 @EntityView(Person.class)
-public interface PersonUpdateView {
+public interface PersonUpdateView extends Identifiable<UUID> {
 
     @IdMapping
+    @Override
     UUID getId();
 
     String getName();
