+++ integration/spring-data/testsuite/webmvc/src/test/java/com/blazebit/persistence/spring/data/testsuite/webmvc/view/PersonView.java
@@ -5,6 +5,7 @@
 
 package com.blazebit.persistence.spring.data.testsuite.webmvc.view;
 
+import com.blazebit.persistence.spring.data.testsuite.webmvc.entity.Identifiable;
 import com.blazebit.persistence.spring.data.testsuite.webmvc.entity.Person;
 import com.blazebit.persistence.view.EntityView;
 import com.blazebit.persistence.view.IdMapping;
@@ -16,9 +17,10 @@ import java.util.UUID;
  * @since 1.5.0
  */
 @EntityView(Person.class)
-public interface PersonView {
+public interface PersonView extends Identifiable<UUID> {
 
     @IdMapping
+    @Override
     UUID getId();
 
     String getName();
